Hackish fix for subseq.
The problem is that when you look for "ids after "test"",
the database finds keys like "test:edges:foo", and thinks it
should include "test" in the results.

I'm currently just appending a ; after the search key in this
case, but this needs a serious fix when we get custom key codecs.