support reading merge-node and unmerge-node in transactions
Add update-wrap-read helper for modifying the last :wrap-read in an
ioval. This is used for both forward-reads and merge-reads now.

Note that merge-reads does not try to make a read function that merges
the given nodes directly- this would make it impossible to unmerge-
instead it delegates to the underlying layer and merge-layer, using the
appropriate logic to merge the results if necessary.