Add a single-type wrapper layer.
It assumes all node-ids start with a given prefix (for example,
"profile-"), and uses 8 bytes to encode the suffix (id number) as a
binary-sortable integer.

Had to update the graph-test to use "real" node-ids like profile-1
instead of the made-up ids like carla, to facilitate running single-type
layers through the tests present in graph-test.