Change how ruminating layers handle max-revision.
If revision 10 was written to the "main" layer, but only revision 9
to the output layer, the ruminating layer should report max-revision
as only 9, because the output needs to be caught up.

However, this means that every revision needs to touch the output
layers in order to get a consistent view of max-revision across all
layers. So, updated the tests to do this as well.