rework how custom key codecs are handled for sorted layers
We cannot fetch everything using ranges, because this means we read
unnecessary data in some cases. For example, if we want to read
["person-1" :age] using the default layout, we end up reading everything
under [:edges :*] even though we don't need it. So now, we group the
subnode-layout by prefix and fetch each group using either a range or a
direct lookup, depending on what is more efficient.